requirements.txt
pandas
numpy
openpyxl
python-dateutil
yfinance==0.2.66
python-certifi-win32